// src/api-chain/model/bankAccountInfo.ts
export interface BankAccountInfo { 
    /**
     * Bank account holder's name
     */
    accountHoldersName?: string;
    /**
     * Bank account number
     */
    accountNumber?: string;
    /**
     * Bank name
     */
    bankName?: string;
    /**
     * Bank's branch address
     */
    branchAddress?: string;
    country?: ApiCountry;
}

/**
 * Namespace for property- and property-value-enumerations of BankAccountInfo.
 */
export namespace BankAccountInfo {
    /**
     * All properties of BankAccountInfo.
     */
    export enum Properties {
        /**
         * Bank account holder's name
         */
        accountHoldersName = 'accountHoldersName',
        /**
         * Bank account number
         */
        accountNumber = 'accountNumber',
        /**
         * Bank name
         */
        bankName = 'bankName',
        /**
         * Bank's branch address
         */
        branchAddress = 'branchAddress',
        country = 'country'
    }


    export function formMetadata() {
        return  {
            metadata: formMetadata,
            classname: 'BankAccountInfo',
            vars: [
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'accountHoldersName',
                    classname: 'BankAccountInfo',
                    dataType: 'string',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'accountNumber',
                    classname: 'BankAccountInfo',
                    dataType: 'string',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'bankName',
                    classname: 'BankAccountInfo',
                    dataType: 'string',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'branchAddress',
                    classname: 'BankAccountInfo',
                    dataType: 'string',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    metadata: ApiCountry.formMetadata,
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'country',
                    classname: 'BankAccountInfo',
                    dataType: 'ApiCountry',
                    isPrimitiveType: false,
                    isListContainer: false,
                    complexType: 'ApiCountry'
                },
            ],
            validators: {
                accountHoldersName: [
                ],
                accountNumber: [
                ],
                bankName: [
                ],
                branchAddress: [
                ],
                country: [
                ],
            }
        }
    }

  // export const BankAccountInfoValidationScheme = {
  //     validators: [],
  //     fields: {
  //               accountHoldersName: {
  //                   validators: []
  //               },
  //               accountNumber: {
  //                   validators: []
  //               },
  //               bankName: {
  //                   validators: []
  //               },
  //               branchAddress: {
  //                   validators: []
  //               },
  //               country: {
  //                   validators: []
  //               },
  //     }
  // } as SimpleValidationScheme<BankAccountInfo>;


}