// src/api-chain/model/farmInfo.ts
export interface FarmInfo { 
    /**
     * Whether a user customer owns a farm.
     */
    ownsFarm?: boolean;
    /**
     * Farm size
     */
    farmSize?: string;
    /**
     * Number of trees
     */
    numberOfTrees?: number;
    /**
     * True if it is an organic farm
     */
    organicFarm?: boolean;
    /**
     * Fertilizer used description.
     */
    fertilizerDescription?: string;
    /**
     * Any additional info.
     */
    additionalInfo?: string;
}

/**
 * Namespace for property- and property-value-enumerations of FarmInfo.
 */
export namespace FarmInfo {
    /**
     * All properties of FarmInfo.
     */
    export enum Properties {
        /**
         * Whether a user customer owns a farm.
         */
        ownsFarm = 'ownsFarm',
        /**
         * Farm size
         */
        farmSize = 'farmSize',
        /**
         * Number of trees
         */
        numberOfTrees = 'numberOfTrees',
        /**
         * True if it is an organic farm
         */
        organicFarm = 'organicFarm',
        /**
         * Fertilizer used description.
         */
        fertilizerDescription = 'fertilizerDescription',
        /**
         * Any additional info.
         */
        additionalInfo = 'additionalInfo'
    }


    export function formMetadata() {
        return  {
            metadata: formMetadata,
            classname: 'FarmInfo',
            vars: [
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'ownsFarm',
                    classname: 'FarmInfo',
                    dataType: 'boolean',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'farmSize',
                    classname: 'FarmInfo',
                    dataType: 'string',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'numberOfTrees',
                    classname: 'FarmInfo',
                    dataType: 'number',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'organicFarm',
                    classname: 'FarmInfo',
                    dataType: 'boolean',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'fertilizerDescription',
                    classname: 'FarmInfo',
                    dataType: 'string',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'additionalInfo',
                    classname: 'FarmInfo',
                    dataType: 'string',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
            ],
            validators: {
                ownsFarm: [
                ],
                farmSize: [
                ],
                numberOfTrees: [
                ],
                organicFarm: [
                ],
                fertilizerDescription: [
                ],
                additionalInfo: [
                ],
            }
        }
    }

  // export const FarmInfoValidationScheme = {
  //     validators: [],
  //     fields: {
  //               ownsFarm: {
  //                   validators: []
  //               },
  //               farmSize: {
  //                   validators: []
  //               },
  //               numberOfTrees: {
  //                   validators: []
  //               },
  //               organicFarm: {
  //                   validators: []
  //               },
  //               fertilizerDescription: {
  //                   validators: []
  //               },
  //               additionalInfo: {
  //                   validators: []
  //               },
  //     }
  // } as SimpleValidationScheme<FarmInfo>;


}